Split utils from helpers as the file was getting big
